+++ app-example/src/test/java/com/badoo/ribs/example/welcome/WelcomeInteractorTest.kt
@@ -1,8 +1,8 @@
 package com.badoo.ribs.example.welcome
 
 import androidx.lifecycle.Lifecycle
-import com.badoo.common.ribs.rx.createInteractorTestHelper
-import com.badoo.common.ribs.rx.mockIO
+import com.badoo.common.ribs.rx2.createInteractorTestHelper
+import com.badoo.common.ribs.rx2.mockIO
 import com.badoo.ribs.example.auth.AuthDataSource
 import com.badoo.ribs.example.welcome.Welcome.Output
 import com.badoo.ribs.example.welcome.WelcomeView.Event
