+++ sandbox/src/test/java/com/badoo/ribs/sandbox/rib/switcher/SwitcherInteractorTest.kt
@@ -2,7 +2,7 @@ package com.badoo.ribs.sandbox.rib.switcher
 
 import androidx.lifecycle.Lifecycle.State.CREATED
 import androidx.lifecycle.Lifecycle.State.STARTED
-import com.badoo.common.ribs.rx.createInteractorTestHelper
+import com.badoo.common.ribs.rx2.createInteractorTestHelper
 import com.badoo.ribs.routing.router.Router.TransitionState.SETTLED
 import com.badoo.ribs.routing.source.backstack.BackStack
 import com.badoo.ribs.routing.source.backstack.operation.push
